Updated packaging to setup.cfg method, various small updates
- Updated the setuptools configuration to use the newer declarative
config method (`setup.cfg` and `pyproject.toml`). `pip install` will
now automatically grab dependencies of the project, and it will be
trivial to upload to PyPI.
- Moved modules into a `archeryutils` directory, allowing a single
entrypoint to the library. The dir `archerycls` was removed, and
the modules inside are now available at the top level of
archeryutils. `handicaps` is now a subpackage of `archeryutils`.
- `handicap_equations` and `handicap_functions` can be imported using
`import archeryutils.handicap_equations` etc. (no need to import
via `handicaps`).
- Some simplification at the end of `rounds.py`
- Simplified `print_handicap_table`: now creates a single string
that can be used either when writing to terminal or writing to
file. Reduces repeat code.
- black formatting (I realise it's doing something awful to the
`arrow_score` equations -- we may need to look at a better way to
achieve those! Maybe it could be a function of `Target`?)
- Updates to keep flake8 happy (shortened long strings/comments,
removed unused imports, etc)