Fixes an issue that resulted in division by zero in the rootfinding algorithm.
This occurred when the score for which a handicap is requested
is the same number of points below the maximum score for a
round as one of the arbitrary bounds used to initialise the
search interval.

Not a bug per-se, as the algorithm is robust enough to handle
division by zero and select a better interval to return the
'correct' handicap, but resulted in a warniong being printed
to the repl or stdout.
This is resolved by replacing the zero with a small number
(xtol already in the function) when it occurs.

Returned handicaps may now be different in this instance by
the 6th sig fig point, but this is within the tolerance that
archeryutils pupports to give and all tests still pass.