Add coaxing functions to classifications.
Captures invalid ages and bowstyles as inputs to classification calculations and coaxes them into a valid format based on ArcheryGB rules. Not applied by default in code but can be used by user to return a dict that can be unpacked directly into the classification functions. Also includes tests for this functionality.